requirements.txt
frappe
facebook-sdk
tweepy
